requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
selenium
webdriver-manager
ultralytics
Pillow
beautifulsoup4
scipy
requests
chromadb
openai
python-dotenv
matplotlib
seaborn
fastapi
uvicorn
python-multipart
huggingface-hub
lightgbm
google-generativeai
